[TensorExpr] Do not inline autodiff graphs if they contain prim::TypeCheck nodes.
Before this change we sometimes inlined autodiff subgraph containing
fusion groups. This happened because we didn't look for 'unsupported'
nodes recursively, but fusion groups were inside if-nodes. With this
change we start to look into sub-blocks and also block-list
prim::TypeCheck - if we were to inline those node we would need to
update the expected types and such logic doesn't exist yet.

The problem was detected by @bertmaher in 'LearningToPaint' benchmark
investigation where this bug caused us to keep constantly hitting
fallback paths of the graph.

ghstack-source-id: 1396f879f2bd7cac8f22b43600e6928330d52667
Pull Request resolved: https://github.com/pytorch/pytorch/pull/44564